moved test_process_and_combine_datasets to test_pvnet_uk_regional.py
