+++ tests/torch_datasets/test_process_and_combine.py
@@ -4,11 +4,7 @@ import xarray as xr
 import dask.array as da
 
 from ocf_data_sampler.config import load_yaml_configuration
-from ocf_data_sampler.select.location import Location
-from ocf_data_sampler.numpy_sample import NWPSampleKey, GSPSampleKey, SatelliteSampleKey
-from ocf_data_sampler.torch_datasets import PVNetUKRegionalDataset
 
-from ocf_data_sampler.torch_datasets.pvnet_uk_regional import process_and_combine_datasets
 
 from ocf_data_sampler.torch_datasets.process_and_combine import (
     process_and_combine_site_sample_dict,
@@ -16,52 +12,6 @@ from ocf_data_sampler.torch_datasets.process_and_combine import (
     fill_nans_in_arrays,
     compute,
 )
-
-
-def test_process_and_combine_datasets(pvnet_config_filename):
-
-    # Load in config for function and define location 
-    config = load_yaml_configuration(pvnet_config_filename)
-    t0 = pd.Timestamp("2024-01-01 00:00")
-    location = Location(coordinate_system="osgb", x=1234, y=5678, id=1)
-
-    nwp_data = xr.DataArray(
-        np.random.rand(4, 2, 2, 2),
-        dims=["time_utc", "channel", "y", "x"],
-        coords={
-            "time_utc": pd.date_range("2024-01-01 00:00", periods=4, freq="h"),
-            "channel": ["t2m", "dswrf"],
-            "step": ("time_utc", pd.timedelta_range(start='0h', periods=4, freq='h')),
-            "init_time_utc": pd.Timestamp("2024-01-01 00:00")
-        }
-    )
-
-    sat_data = xr.DataArray(
-        np.random.rand(7, 1, 2, 2),
-        dims=["time_utc", "channel", "y", "x"],
-        coords={
-            "time_utc": pd.date_range("2024-01-01 00:00", periods=7, freq="5min"),
-            "channel": ["HRV"],
-            "x_geostationary": (["y", "x"], np.array([[1, 2], [1, 2]])),
-            "y_geostationary": (["y", "x"], np.array([[1, 1], [2, 2]]))
-        }
-    )
-
-    # Combine as dict
-    dataset_dict = {
-        "nwp": {"ukv": nwp_data},
-        "sat": sat_data
-    }
-
-    # Call relevant function
-    result = process_and_combine_datasets(dataset_dict, config, t0, location)
-
-    # Assert result is dict - check and validate
-    assert isinstance(result, dict)
-    assert NWPSampleKey.nwp in result
-    assert result[SatelliteSampleKey.satellite_actual].shape == (7, 1, 2, 2)
-    assert result[NWPSampleKey.nwp]["ukv"][NWPSampleKey.nwp].shape == (4, 1, 2, 2)
-
 
 def test_merge_dicts():
     """Test merge_dicts function"""
